requirements.txt
chromadb>=0.4.24
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
